+++ src/Traits/Configuration/SearchConfiguration.php
@@ -149,4 +149,11 @@ trait SearchConfiguration
 
         return $this;
     }
+
+    public function setSearchFieldAttributes(array $attributes = []): self
+    {
+        $this->searchFieldAttributes = $attributes;
+
+        return $this;
+    }
 }